# Use PyTorch CUDA memory allocator to manage small memory allocation for batch inputs