This lets us remove repeated `cudaMalloc` of very small GPU allocations, that need to happen at every decoding step.   

7B fp16
```
Before: Throughput: 7.17 requests/s, 3427.27 tokens/s
After: Throughput: 7.26 requests/s, 3473.79 tokens/s
```

13B fp16
```
Before: Throughput: 3.84 requests/s, 1834.09 tokens/s
After: Throughput: 3.89 requests/s, 1858.39 tokens/s
```